src/ai-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
pydantic==2.9.2
scikit-learn==1.5.2
pandas==2.2.3
numpy==1.26.4
scipy==1.14.1
pymongo==4.10.1
python-dotenv==1.0.1
httpx==0.27.2
joblib==1.4.2
